Rename encryption_common module to encryption_agnostic
