+++ parquet/tests/arrow_reader/mod.rs
@@ -40,9 +40,9 @@ mod bad_data;
 mod checksum;
 #[cfg(feature = "encryption")]
 mod encryption;
+mod encryption_agnostic;
 #[cfg(all(feature = "encryption", feature = "async"))]
 mod encryption_async;
-mod encryption_common;
 #[cfg(not(feature = "encryption"))]
 mod encryption_disabled;
 #[cfg(feature = "encryption")]
